+++ src/main/resources/static/js/admin.js
@@ -48,7 +48,7 @@ form.addEventListener('submit', (event) => {
 const createProduct = (product) => {
     console.log(product);
     axios.request({
-        url: '/admin/products',
+        url: '/products',
         method: 'post',
         data: product
     }).then((response) => {
@@ -64,7 +64,7 @@ const updateProduct = (product) => {
     const {id} = product;
 
     axios.request({
-        url: '/admin/products/' + id,
+        url: '/products/' + id,
         method: 'put',
         data: product
     }).then((response) => {
@@ -78,7 +78,7 @@ const updateProduct = (product) => {
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const deleteProduct = (id) => {
     axios.request({
-        url: '/admin/products/' + id,
+        url: '/products/' + id,
         method: 'delete'
     }).then((response) => {
         window.location.reload();
