+++ src/test/java/cart/ProductIntegrationTest.java
@@ -27,7 +27,7 @@ public class ProductIntegrationTest {
         var result = given()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .get("/")
+                .get("/products")
                 .then()
                 .extract();
