+++ src/test/java/cart/controller/MainControllerTest.java
@@ -27,5 +27,4 @@ class MainControllerTest {
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value());
     }
-
 }
